changed int to float in default

+++ ex2/main.py
@@ -34,7 +34,7 @@ PSEUDO_WORDS_PATTERNS = {
 OTHER_WORDS = 'OtherCategory'
 
 def default():
-    return defaultdict(int)
+    return defaultdict(float)
 
 
 class BasicModel:
